修改好友获取巨大量数据的内存问题， 同步周俊的代码
修改好友获取巨大量数据的内存问题， 同步周俊的代码

+++ platforms/ios/bsl/Classes/iPad/MainView/controller/MainViewViewController.m
@@ -298,7 +298,7 @@
                     while([result next])
                     {
                         NSString * showTimeTmp = [result objectForColumnName:@"showTime"];
-                        long showTime = [showTimeTmp longLongValue];
+                        long long showTime = [showTimeTmp longLongValue];
                         
                         if([module.timeUnit isEqualToString:@"H"])
                         {
